Initial commit
Far from complete, but on the right track well enough to start
versioning my Vim customization.

I'm using vim-plug to manage plugins. vimrc contains a few
lines to download vim-plug via curl if it doesn't exist yet in the
expected location, which is ~/.local/share/vim/autoload.

The idea is that this repo exclusively contains files that I
have directly introduced for the sake of customization. Any files
that are downloaded, such as vim-plug itself and the plugins it
downloads, as well as generated while using Vim, such as swap
files and command history, are not in this repo.

Note that the above exclusion is implemented by specifying non-default
directories for said files. So far, no .gitignore is needed, and the
plan is to keep it this way.

I'm still a bit on the fence regarding the exclusion of history, simply
because it's often quite convenient and useful to have a bunch of
things that I regularly reach for in the history. At least, that's
my experience with shells. On the other hand, if I transfer my
configuration to another machine, chances are that the history contains
references to projects that do not exist on that machine.